+++ packages/core/src/common/number/view/slider.ts
@@ -1,9 +1,9 @@
 import {Value} from '../../../common/model/value';
 import {ValueMap} from '../../../common/model/value-map';
 import {ViewProps} from '../../../common/model/view-props';
-import {constrainRange, mapRange} from '../../../common/number-util';
 import {ClassName} from '../../../common/view/class-name';
 import {View} from '../../../common/view/view';
+import {constrainRange, mapRange} from '../util';
 
 /**
  * @hidden
