+++ packages/core/src/index.ts
@@ -101,11 +101,11 @@ export * from './common/view/reactive';
 export * from './common/view/text';
 export * from './common/view/view';
 
+export * from './common/number/util';
 export * from './common/compat';
 export * from './common/dom-util';
 export * from './common/list-util';
 export * from './common/micro-parsers';
-export * from './common/number-util';
 export * from './common/params';
 export * from './common/picker-util';
 export * from './common/point-nd-util';
