+++ packages/core/src/input-binding/color/view/sv-palette.ts
@@ -1,7 +1,7 @@
 import {getCanvasContext} from '../../../common/dom-util';
 import {Value} from '../../../common/model/value';
 import {ViewProps} from '../../../common/model/view-props';
-import {mapRange} from '../../../common/number-util';
+import {mapRange} from '../../../common/number/util';
 import {ClassName} from '../../../common/view/class-name';
 import {View} from '../../../common/view/view';
 import {hsvToRgbInt} from '../model/color-model';
