+++ packages/core/src/input-binding/point-2d/plugin-test.ts
@@ -11,9 +11,9 @@ import {StepConstraint} from '../../common/constraint/step';
 import {ComplexValue} from '../../common/model/complex-value';
 import {getBoundValue} from '../../common/model/test-util';
 import {Value} from '../../common/model/value';
+import {findNumberRange} from '../../common/number/util';
 import {createTestWindow} from '../../misc/dom-test-util';
 import {PointNdConstraint} from '../common/constraint/point-nd';
-import {findNumberRange} from '../number/plugin';
 import {createInputBindingController} from '../plugin';
 import {Point2d, Point2dAssembly} from './model/point-2d';
 import {getSuitableMax, Point2dInputPlugin} from './plugin';
