+++ packages/core/src/input-binding/point-3d/plugin-test.ts
@@ -7,10 +7,10 @@ import {findConstraint} from '../../common/constraint/composite';
 import {StepConstraint} from '../../common/constraint/step';
 import {ComplexValue} from '../../common/model/complex-value';
 import {getBoundValue} from '../../common/model/test-util';
+import {findNumberRange} from '../../common/number/util';
 import {createTestWindow} from '../../misc/dom-test-util';
 import {PointNdConstraint} from '../common/constraint/point-nd';
 import {PointNdTextController} from '../common/controller/point-nd-text';
-import {findNumberRange} from '../number/plugin';
 import {createInputBindingController} from '../plugin';
 import {Point3d} from './model/point-3d';
 import {Point3dInputPlugin} from './plugin';
