+++ packages/core/src/input-binding/point-4d/plugin.ts
@@ -9,15 +9,17 @@ import {
 	getBaseStep,
 	getSuitableDecimalDigits,
 	getSuitableDraggingScale,
-} from '../../common/number-util';
+} from '../../common/number/util';
 import {BaseInputParams, PointDimensionParams} from '../../common/params';
-import {parsePointDimensionParams} from '../../common/point-nd-util';
+import {
+	createDimensionConstraint,
+	parsePointDimensionParams,
+} from '../../common/point-nd-util';
 import {TpError} from '../../common/tp-error';
 import {VERSION} from '../../version';
 import {PointNdConstraint} from '../common/constraint/point-nd';
 import {PointNdTextController} from '../common/controller/point-nd-text';
 import {InputBindingPlugin} from '../plugin';
-import {createDimensionConstraint} from '../point-2d/plugin';
 import {point4dFromUnknown, writePoint4d} from './converter/point-4d';
 import {Point4d, Point4dAssembly, Point4dObject} from './model/point-4d';
